Saving, Goals and Imports
Added in initial code to EntityList that allows it to jsonify its entities for saving. Ran into issue when trying to jsonify goal code, and generalised the code to work better with saving/loading, but to be more extensible overall. This was done by using the component pattern. Removed some redundant imports as well