Spider Reverse, Rare Spawn/Images, Debug Prints,
Spiders now reverse their direction in the same way as the example code when they hit a boulder.
Fixed a bug in rare item spawning where the rare item would be on the map and in the players inventory simultaneously.
Removed white background from one ring and removed anduril background to fit the art style.
Removed redstone related debug prints.